requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
torch==2.1.0
torchvision==0.16.0
diffusers==0.25.0
transformers==4.35.0
accelerate==0.25.0
safetensors==0.4.1
opencv-python-headless==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
python-dotenv==1.0.0
gunicorn==21.2.0
peft==0.7.0
scikit-image==0.22.0
